chore: add few todos and moved classes to seperate packages
