+++ src/main/java/com/vi/appointmentservice/controller/AskerController.java
@@ -7,8 +7,8 @@ import com.vi.appointmentservice.generated.api.controller.AskersApi;
 import com.vi.appointmentservice.helper.RescheduleHelper;
 import com.vi.appointmentservice.model.CalcomBookingToAsker;
 import com.vi.appointmentservice.repository.CalcomBookingToAskerRepository;
-import com.vi.appointmentservice.service.CalComBookingService;
-import com.vi.appointmentservice.service.CalcomRepository;
+import com.vi.appointmentservice.service.calcom.CalComBookingService;
+import com.vi.appointmentservice.repository.CalcomRepository;
 import com.vi.appointmentservice.service.CalcomWebhookHandlerService;
 import io.swagger.annotations.Api;
 import java.util.ArrayList;
