+++ src/main/java/com/vi/appointmentservice/helper/RescheduleHelper.java
@@ -5,9 +5,9 @@ import com.vi.appointmentservice.api.model.CalcomBooking;
 import com.vi.appointmentservice.api.model.CalcomUser;
 import com.vi.appointmentservice.repository.CalcomBookingToAskerRepository;
 import com.vi.appointmentservice.repository.CalcomUserToConsultantRepository;
-import com.vi.appointmentservice.service.AdminUserService;
-import com.vi.appointmentservice.service.CalComEventTypeService;
-import com.vi.appointmentservice.service.CalComUserService;
+import com.vi.appointmentservice.service.onlineberatung.AdminUserService;
+import com.vi.appointmentservice.service.calcom.CalComEventTypeService;
+import com.vi.appointmentservice.service.calcom.CalComUserService;
 import com.vi.appointmentservice.useradminservice.generated.web.model.AskerResponseDTO;
 import com.vi.appointmentservice.useradminservice.generated.web.model.ConsultantDTO;
 import lombok.AllArgsConstructor;
