+++ src/main/java/com/vi/appointmentservice/service/onlineberatung/AdminUserService.java
@@ -1,4 +1,4 @@
-package com.vi.appointmentservice.service;
+package com.vi.appointmentservice.service.onlineberatung;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
@@ -21,7 +21,7 @@ import org.springframework.stereotype.Service;
 @Service
 @Slf4j
 //TODO: what do you think that we move all onlineberatung services like, user, useradmin, messages
-// to one subpacke of service package
+// to one subpackage of service package
 //TODO: and all calcom services to calcom package. classes would not have calcom prefix than
 public class AdminUserService {
