+++ src/main/java/com/vi/appointmentservice/service/onlineberatung/MessagesService.java
@@ -1,4 +1,4 @@
-package com.vi.appointmentservice.service;
+package com.vi.appointmentservice.service.onlineberatung;
 
 import com.vi.appointmentservice.adapters.keycloak.dto.KeycloakLoginResponseDTO;
 import com.vi.appointmentservice.api.model.CalcomBooking;
@@ -10,6 +10,7 @@ import com.vi.appointmentservice.model.CalcomUserToConsultant;
 import com.vi.appointmentservice.port.out.IdentityClient;
 import com.vi.appointmentservice.repository.CalcomBookingToAskerRepository;
 import com.vi.appointmentservice.repository.CalcomUserToConsultantRepository;
+import com.vi.appointmentservice.service.calcom.CalComBookingService;
 import com.vi.appointmentservice.service.securityheader.SecurityHeaderSupplier;
 import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
